+++ woocommerce/wc-functions.php
@@ -33,7 +33,8 @@ require_once('inc/blocks/wc-block-widget-categories.php');
  * Register Ajax Cart
  *
  * Enabled/Disabled based on the setting in backend under WooCommerce > Settings > Products > Enable AJAX add to cart buttons on archives.
- * Disable file via filter add_filter('bootscore/load_ajax_cart', '__return_false');
+ * Disable file via filter 
+ * add_filter('bootscore/load_ajax_cart', '__return_false');
  */
 function bootscore_register_ajax_cart() {
   if (apply_filters('bootscore/load_ajax_cart', true)) {
@@ -44,3 +45,22 @@ function bootscore_register_ajax_cart() {
   }
 }
 add_action('after_setup_theme', 'bootscore_register_ajax_cart');
+
+
+/**
+ * Skip cart page
+ *
+ * Disable cart page, "View Cart" button in mini-cart and redirect cart page to checkout
+ *
+ * Enable default cart page and buttoins via filter 
+ * add_filter('bootscore/skip_cart', '__return_false');
+ */
+function bootscore_register_cart_file() {
+  if (apply_filters('bootscore/skip_cart', true)) {
+    require_once('inc/wc-skip-cart.php');
+  } else {
+    require_once('inc/wc-cart.php');
+  }
+}
+add_action('after_setup_theme', 'bootscore_register_cart_file');
+
